Removing page titles from configuration
I wasn't comfortable with them define in there anyway. It's better to
include the header file from inside the individual pages and define the
title variables before that. This way titles and any other information
will be available inside the contents pages. Easier to manage and keep up
to date.